:broom: moved faq into the separate component
